# docs/swagger.yaml
definitions:
  apihelpers.APIRes:
    properties:
      data: {}
      message:
        type: string
      status:
        type: boolean
    type: object
  models.Login:
    properties:
      emailId:
        type: string
      password:
        type: string
    type: object
  models.SignUp:
    properties:
      emailId:
        type: string
      name:
        type: string
      password:
        type: string
      phoneNumber:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/auth/login:
    post:
      description: API for user login
      parameters:
      - description: Login Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Login'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apihelpers.APIRes'
      summary: User Login
      tags:
      - Auth
  /v1/auth/send-email-otp:
    get:
      description: API for sending otp by email
      parameters:
      - default: satish.sund3r@gmail.com
        description: email
        in: query
        name: email
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apihelpers.APIRes'
      summary: send email otp
      tags:
      - Auth
  /v1/auth/signup:
    post:
      description: API for user signup
      parameters:
      - description: Signup Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SignUp'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apihelpers.APIRes'
      summary: User Signup
      tags:
      - Auth
  /v1/auth/verify-email-otp:
    get:
      description: API for verifying otp by email
      parameters:
      - default: satish.sund3r@gmail.com
        description: email
        in: query
        name: email
        required: true
        type: string
      - description: otp
        in: query
        name: otp
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apihelpers.APIRes'
      summary: verify email otp
      tags:
      - Auth
swagger: "2.0"