# router/routers.go
package router

import (
	"auth-service/controller"

	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
)

func InitRouters() *gin.Engine {
	r := gin.Default()

	r.GET("/swagger/auth-service/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
	r.Use(func(c *gin.Context) {
		// add header Access-Control-Allow-Origin
		c.Writer.Header().Set("Content-Type", "application/json")
		c.Writer.Header().Set("Access-Control-Allow-Origin", "*")
		c.Writer.Header().Set("Access-Control-Max-Age", "86400")
		c.Writer.Header().Set("Access-Control-Allow-Methods", "POST, GET, PUT, DELETE, UPDATE")
		c.Writer.Header().Set("Access-Control-Allow-Headers", "Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, X-Max")
		c.Writer.Header().Set("Access-Control-Allow-Credentials", "true")

		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(200)
		} else {
			c.Next()
		}
	})

	v1Auth := r.Group("/v1/auth")
	{
		v1Auth.POST("/signup", controller.SignUp)
		v1Auth.GET("/send-email-otp", controller.SendEmailOTP)
		v1Auth.GET("/verify-email-otp", controller.VerifyEmailOtp)
		v1Auth.POST("/login", controller.Login)
	}
	return r
}